CDSR-1707: add routes and move controller
