+++ conf/overpayments_scheduled.routes
@@ -18,6 +18,12 @@ POST        /scheduled/claimant-details                                         
 GET         /scheduled/claimant-details/lookup-address                          @uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.overpaymentsscheduled.CheckContactDetailsController.redirectToALF
 GET         /scheduled/claimant-details/update-address                          @uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.overpaymentsscheduled.CheckContactDetailsController.retrieveAddressFromALF(id: Option[UUID] ?= None)
 
+GET         /scheduled/claimant-details/change-contact-details                  @uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.overpaymentsscheduled.EnterContactDetailsController.changeContactDetails
+POST        /scheduled/claimant-details/change-contact-details                  @uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.overpaymentsscheduled.EnterContactDetailsController.changeContactDetailsSubmit
+
+GET         /scheduled/claimant-details/enter-contact-details                   @uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.overpaymentsscheduled.EnterContactDetailsController.enterContactDetails
+POST        /scheduled/claimant-details/enter-contact-details                   @uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.overpaymentsscheduled.EnterContactDetailsController.enterContactDetailsSubmit
+
 GET         /scheduled/claim-northern-ireland                                   @uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.overpaymentsscheduled.NorthernIrelandController.show
 POST        /scheduled/claim-northern-ireland                                   @uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.overpaymentsscheduled.NorthernIrelandController.submit
